move blur check to content component - closes #297
