+++ src/components/leaf.js
@@ -6,24 +6,6 @@ import ReactDOM from 'react-dom'
 import getWindow from 'get-window'
 
 /**
- * Tests if the child is a descendant of parent.
- *
- * @param parent
- * @param child
- * @returns {boolean}
- */
-
-function isDescendant(parent, child) {
-  if (!child || !parent) return false
-  let node = child.parentNode
-  while (node != null) {
-    if (node == parent) return true
-    node = node.parentNode
-  }
-  return false
-}
-
-/**
  * Debugger.
  *
  * @type {Function}
@@ -154,6 +136,9 @@ class Leaf extends React.Component {
     const { state, ranges, isVoid } = this.props
     const { selection } = state
 
+    // If the selection is blurred we have nothing to do.
+    if (selection.isBlurred) return
+
     let { anchorOffset, focusOffset } = selection
     const { node, index } = this.props
     const { start, end } = OffsetKey.findBounds(index, ranges)
@@ -163,42 +148,19 @@ class Leaf extends React.Component {
     const hasFocus = selection.hasFocusBetween(node, start, end)
     if (!hasAnchor && !hasFocus) return
 
-    // We have a selection to render, so prepare a few things...
-    const ref = ReactDOM.findDOMNode(this)
-    const el = findDeepestNode(ref)
-    const window = getWindow(el)
-
-    // If no selection is defined we're probably in a node environment and we can skip the selection magic.
-    if (!window || !window.getSelection) return
-
-    const native = window.getSelection()
-    const parent = ref.closest('[contenteditable]')
-
-    // If the selection is blurred but the DOM selection still focuses this leaf,
-    // we need to clean up the ranges and blur the contenteditable.
-    if (selection.isBlurred && (hasAnchor || hasFocus)) {
-      // We need to make sure that the selection from our state is up-to-date with the native selection.
-      // We can do this by checking if native.anchorNode is a descendant of our contenteditable parent, which is this
-      // slate instance. If it's not, the selection is no longer under this instance's responsibility and not further
-      // action is required.
-      if (!isDescendant(parent, native.anchorNode)) return
-
-      // Apparently our selection is blurred, but the DOM still has ranges in the nodes we manage. To fix this,
-      // we blur the contenteditable and remove all ranges.
-      native.removeAllRanges()
-      if (parent) parent.blur()
-      return
-    }
-
-    // If the selection is blurred we have nothing to do.
-    if (selection.isBlurred) return
-
     // If the leaf is a void leaf, ensure that it has no width. This is due to
     // void nodes always rendering an empty leaf, for browser compatibility.
     if (isVoid) {
       anchorOffset = 0
       focusOffset = 0
     }
+
+    // We have a selection to render, so prepare a few things...
+    const ref = ReactDOM.findDOMNode(this)
+    const el = findDeepestNode(ref)
+    const window = getWindow(el)
+    const native = window.getSelection()
+    const parent = ref.closest('[contenteditable]')
 
     // In firefox it is not enough to create a range, you also need to focus the contenteditable element.
     function focus() {
